Start using Heroku-hosted Node.js API as source of school data
Prior to this change, there were attempts to access school data directly
from mLab using MongoDB driver for C#. The attempts failed.

This change will instead use the Node.js app hosted on Heroku as the
source of data. The Node.js app uses mongodb npm package to access data
from the database.

Changes were made to models to work with the new structure. There is a
SchoolService class that fetches the data from the API. Instances of
Query and SchoolService classes are added as singleton via dependency
injection in the Startup class. This change allows the app to fetch
school data successfully.