Adjust the pages and pagemodels for production
Prior to this change, the pages were not properly adjusted to go to
production. The styling was a mess, the About and Contact pages had
wrong/unorganized information. The PageModels did not have necessary
properties to work with the new method of delivery of school data.

This change applies BootStrap to functional pages (Index, Inputs,
Outputs). The data in the informational pages (About, Contact) is also
edited as needed. PageModels are adjusted to work with the new Query
 and SchoolService singletons. Unnecessary properties of PageModels are
also removed.