Add unit tests for models, add test commands to CI/CD scripts
Prior to this change, there was an xUnit test project in the tests
folder, but there was no unit test there. Hence, any mention of tests
were absent in the CI/CD pipeline commands.

This change adds 4 unit tests to that project. All 4 of them are
regarding the Stat class. The test commands are also added to the CI/CD
scripts.