Add Travis CI config
Prior to this change, there was no Travis CI
configuration. Initially this project had Travis
CI configuration, but it was omitted due to
Travis' lack of support for dotnet core 2.2.

This change adds a configuration for Travis
CI, upon finding an apparently reliable CI
configuration on a post on Travis CI
community.

Link: https://travis-ci.community/t/dotnet-core-2-2/1216/12