Add tests, fix FAHP logic
Prior to this change, there was no test for
invalid inputs for Inference and FAHP comparison
matrix. Also, there were two unnecessary
condition checks.

This change adds two tests specifically for
invalid inputs of numbers and arrays. This change
also removes the unnecessary condition checks.

Additionally, this change adds a global.json file
so that the correct version of dotnet core can be
used with this project.