Switch from a class to an interface
Prior to this change, the data service class (SchoolService) was a
concrete class, which made both this class and the Query class (which is
dependent on that class) hard to test.

This change opts out of a concrete class for the service object and
instead goes for an interface, which can easily be tested without
having to implement a functioning data source.

This implements Dependency Injection (DI) principle, the 'D' of 'SOLID'
principle.