Add a test suite for Query class with a unit test
Prior to this change, there was no test for the Query class.

This change adds a suite of unit tests for this class. This suite now
contains one unit test, only testing the Ctor of the class.

There is also a class that implements ISchoolService. This class is used
to mock a data service class that is required for the constructor of the
Query class.