chore: Update readme with badges and minor fixes
Prior to this change, build status badges for Travis CI and AppVeyor
were not shown in readme file, although this project was being
built in those services with each commit. Moreover, the description of
this project was not quite right, there were some links missing in TOC,
and there was a mistaken capitalization of a letter.

This change fixes (almost) all of these issues.