Add environment-based dependency injection
Prior to this change, the only `ISchoolService` implementation that was
injected in Startup is the default `SchoolService` class. This was for
production. While development, this took a long time to load, which was
a dilemma when testing the app on localhost.

This change makes the dependency injection conditional on environment
and brings the `SchoolServiceMock` class to the main project, so that
the right implementation of `ISchoolService` can be injected based on
the environment the app is running in.