Use in-browser session storage instead of DI
Prior to this change, a singleton `Query` object was being used to store
the user's inputted data (i.e. age, class, sex, comparison). The problem
was that the object being a singleton, all the users had the same copy
of that object, hence the same data!

For example, if someone gives a particular set of inputs, the schools
he/she will see along with their compatibility scores can be seen by
anyone by just navigating to the `Outputs` page without having to give
the same inputs!

This change replaces the singleton system with the browser's session
storage. The application reads from and writes in the browser's session
storage when necessary. This way, everyone's data stays in their own
devices.

However, the task is not over yet. Certain measures have to be set so
that nobody can navigate to the `Inputs` or `Outputs` page without
gicing the inputs in the `Index` page. Some unit tests are also in need
of correction.