Move Docker-related files to project root dir
Prior to this change, the two Docker-related files, `Dockerfile` and
`.dockerignore`, were inside `src` folder. This was causing no problem
for building and running locally, and manually publishing images to
Docker Hub was also quite possible. But it was recently discovered that
Docker Hub was not being able to build Docker images for every push to
this repository.

This change moves the files to the root directory of the repo and makes
necessary changes in the file paths of the `Dockerfile`. Now Docker
images can be built both manually by a user on a local machine and on
Docker Hub for every push.