Add support for directly accessing MongoDB Atlas
Prior to this change, the responsibility of accessing the database was
carried out by a Node.js application deployed as a ReST API on Heroku
(see my project `konschool-mlab-host` on GitHub). It was set up in this
way due to issues with the old MongoDB driver and/or mLab. With this set
up, `SchoolService` would make a GET request to that app to get all the
details of the schools (responses were parsed with `Newtonsoft.Json`).

This change implements `SchoolService` with MongoDB driver, so that this
app can connect directly with MongoDB Atlas (switched from mLab), with a
few environment variables.

This commit undoes part of what 50f8f8e had done.