Implement global app state
The state container, `AppStateContainer`, is a refactored version of the
previously used `Query` class. To reduce over-abundance of atomic
variables in this class, `Location` and `Student` classes were created.

The implementation is based on dependency injection. The state container
class is injected in `Startup.ConfigureServices` as a `scoped` service,
so that each client gets their own instance, which is the same one
across all the pages and components for that particular client.

Acknowledgement: https://www.youtube.com/watch?v=GIupo55GTro.