Redesign and refactor Comparison component
The component now uses the `MudSlider` component, along with a few other
MudBlazor components. This change is to make this component visually
coherent with the rest of the application.

The component is also slightly refactored. Instead of getting the whole
app-state-container injected into the component, an `int[]` is passed on
as a parameter. This way, the component becomes more reusable.