Add conditional component for `GetSchoolsError`
Now, if the weights are not fetched, it will show that it cannot show
schools without them. Otherwise, if the weights are loaded but the
schools are not fetched for some reason, it will show that schools could
not be fetched.

In fact, it will not try to fetch schools if the weights could not be
fetched.