Upgrade FAHP service to .NET 6
This project now uses .NET 6 preview 4. It is written as a minimal API
from the empty web project template. Improvement in performance can be
expected, according to Daniel Roth's post on Microsoft Developer Blogs.

The dependency injection in the minimal API structure is yet not quite
clear to me, which led me to use Serilog for logging (and find out how
simpler it is!).