Upgrade the FAHP service to .NET 6 preview 7
The changes from preview 4 come mainly from non-nullable references:
route values are checked before the `try`-`catch` block in the API
delegate method, and all the properties of `FAHP` class are set inside
the `ctor`.

The top-level namespace declaration is used in all the files.

This commit also contains minor upgrades to the `Serilog` packages.